+++ guppylang-internals/src/guppylang_internals/definition/struct.py
@@ -65,6 +65,7 @@ class RawStructDef(TypeDef, ParsableDef, UserProvidedLinkName):
     """A raw struct type definition that has not been parsed yet."""
 
     python_class: type
+    frozen: bool
     params: None = field(default=None, init=False)  # Params not known yet
 
     def parse(self, globals: Globals, sources: SourceMap) -> "ParsedStructDef":
@@ -139,7 +140,7 @@ class RawStructDef(TypeDef, ParsableDef, UserProvidedLinkName):
         )
 
         return ParsedStructDef(
-            self.id, self.name, cls_def, params, fields, link_name_prefix
+            self.id, self.name, cls_def, params, fields, self.frozen, link_name_prefix
         )
 
     def check_instantiate(
@@ -155,6 +156,7 @@ class ParsedStructDef(TypeDef, CheckableDef):
     defined_at: ast.ClassDef
     params: Sequence[Parameter]
     fields: Sequence[UncheckedField]
+    frozen: bool
     link_name_prefix: str
 
     def check(self, globals: Globals) -> "CheckedStructDef":
@@ -170,7 +172,7 @@ class ParsedStructDef(TypeDef, CheckableDef):
             CheckedField(f.name, type_from_ast(f.type_ast, ctx)) for f in self.fields
         ]
         return CheckedStructDef(
-            self.id, self.name, self.defined_at, self.params, fields
+            self.id, self.name, self.defined_at, self.params, fields, self.frozen
         )
 
     def check_instantiate(
@@ -197,6 +199,7 @@ class CheckedStructDef(TypeDef, CompiledDef):
     defined_at: ast.ClassDef
     params: Sequence[Parameter]
     fields: Sequence[CheckedField]
+    frozen: bool
 
     def check_instantiate(
         self, args: Sequence[Argument], loc: AstNode | None = None