+++ guppylang-internals/src/guppylang_internals/nodes.py
@@ -839,8 +839,6 @@ class Modifiers:
             result |= UnitaryFlags.Dagger
         if self.has_control():
             result |= UnitaryFlags.Control
-        if self.has_power():
-            result |= UnitaryFlags.Power
         return result