+++ guppylang/src/guppylang/decorator.py
@@ -94,9 +94,8 @@ class GuppyKwargs(TypedDict, total=False):
     """
 
     unitary: bool
-    control: bool
-    dagger: bool
-    power: bool
+    controllable: bool
+    daggerable: bool
     max_qubits: int
     link_name: str
 
@@ -106,6 +105,7 @@ class GuppyStructKwargs(TypedDict, total=False):
     `@guppy.struct` decorator.
     """
 
+    frozen: bool
     link_name: str
 
 
@@ -250,6 +250,7 @@ class _Guppy:
                 cls.__name__,
                 None,
                 cls,
+                frozen=kwargs.pop("frozen", False),  # Mutable by default
                 link_name=kwargs.pop("link_name", None),
             )
             frame = get_calling_frame()
@@ -833,17 +834,18 @@ def _parse_kwargs(kwargs: GuppyKwargs) -> ParsedGuppyKwargs:
     """Parses the kwargs dict specified in the `@guppy` decorator into `UnitaryFlags`
     and other metadata that will be passed onto the compiled function as is.
     """
+    metadata = FunctionMetadata()
+
     flags = UnitaryFlags.NoFlags
     if kwargs.pop("unitary", False):
         flags |= UnitaryFlags.Unitary
-    if kwargs.pop("control", False):
+    if kwargs.pop("controllable", False):
         flags |= UnitaryFlags.Control
-    if kwargs.pop("dagger", False):
+    if kwargs.pop("daggerable", False):
         flags |= UnitaryFlags.Dagger
-    if kwargs.pop("power", False):
-        flags |= UnitaryFlags.Power
 
-    metadata = FunctionMetadata()
+    metadata.set_unitary_flags(flags.value)
+
     if "max_qubits" in kwargs:
         metadata.set_max_qubits(kwargs.pop("max_qubits"))