+++ tests/error/modifier_errors/bad_custom_call.py
@@ -1,10 +1,10 @@
 from guppylang.decorator import guppy
 from guppylang.std.builtins import owned
-from guppylang.std.qsystem import measure
+from guppylang.std.qsystem.helios import measure
 from guppylang.std.quantum import qubit
 
 
-@guppy(dagger=True)
+@guppy(daggerable=True)
 def test(x: qubit @owned) -> None:
     measure(x)