+++ tests/error/modifier_errors/flag_loop1.py
@@ -1,7 +1,7 @@
 from guppylang.decorator import guppy
 
 
-@guppy(dagger=True)
+@guppy(daggerable=True)
 def test() -> None:
     for _ in range(10):
         pass