+++ tests/integration/test_struct.py
@@ -1,6 +1,9 @@
 from typing import Generic, TYPE_CHECKING
 
+from guppylang import comptime, qubit, array
 from guppylang.decorator import guppy
+from guppylang.std.quantum import discard_array
+
 from tests.integration.modules import struct_scope_defs
 
 
@@ -41,11 +44,11 @@ def test_basic_defs(validate):
 
 
 def test_backward_ref(validate):
-    @guppy.struct
+    @guppy.struct(frozen=True)
     class StructA:
         x: int
 
-    @guppy.struct
+    @guppy.struct(frozen=True)
     class StructB:
         y: StructA
 
@@ -57,11 +60,11 @@ def test_backward_ref(validate):
 
 
 def test_forward_ref(validate):
-    @guppy.struct
+    @guppy.struct(frozen=True)
     class StructA:
         x: "StructB"
 
-    @guppy.struct
+    @guppy.struct(frozen=True)
     class StructB:
         y: int
 
@@ -88,17 +91,17 @@ def test_generic(validate):
     S = guppy.type_var("S")
     T = guppy.type_var("T")
 
-    @guppy.struct
+    @guppy.struct(frozen=True)
     class StructA(Generic[T]):
         x: tuple[int, T]
 
-    @guppy.struct
+    @guppy.struct(frozen=True)
     class StructC:
         a: StructA[int]
         b: StructA[list[bool]]
         c: "StructB[float, StructB[bool, int]]"
 
-    @guppy.struct
+    @guppy.struct(frozen=True)
     class StructB(Generic[S, T]):
         x: S
         y: StructA[T]
@@ -205,3 +208,28 @@ def test_redefine(validate):
         return MyStruct()
 
     validate(foo.compile_function())
+
+
+def test_comptime_attribute_access(validate):
+    """Tests that comptime expressions struct definitions can access attributes of local
+    variables. Regression test for https://github.com/Quantinuum/guppylang/issues/1816.
+    """
+
+    class Config:
+        x: int
+
+        def __init__(self, x: int):
+            self.x = x
+
+    config = Config(x=5)
+
+    @guppy.struct
+    class Struct:
+        qs: array[qubit, comptime(config.x)]
+
+    @guppy
+    def main() -> None:
+        s = Struct(array(qubit() for _ in range(comptime(config.x))))
+        discard_array(s.qs)
+
+    validate(main.compile_function())