+++ guppylang-internals/CHANGELOG.md
@@ -3,6 +3,71 @@
 First release of `guppylang_internals` package containing refactored out internal components
 from `guppylang`.
 
+## [1.0.0-a4](https://github.com/Quantinuum/guppylang/compare/guppylang-internals-v1.0.0-a3...guppylang-internals-v1.0.0-a4) (2026-06-12)
+
+
+### ⚠ BREAKING CHANGES
+
+* Change the return type of measurement functions to a new `Measurement` type and remove opaque bools from the compiler ([#1707](https://github.com/Quantinuum/guppylang/issues/1707))
+* Replace track_hugr_side_effects monkeypatch with hierarchy of builder classes ([#1829](https://github.com/Quantinuum/guppylang/issues/1829))
+
+### Features
+
+* Change the return type of measurement functions to a new `Measurement` type and remove opaque bools from the compiler ([#1707](https://github.com/Quantinuum/guppylang/issues/1707)) ([9460e6b](https://github.com/Quantinuum/guppylang/commit/9460e6b672fe84d87f360b7d521b6b9d5b736da2))
+
+
+### Code Refactoring
+
+* Replace track_hugr_side_effects monkeypatch with hierarchy of builder classes ([#1829](https://github.com/Quantinuum/guppylang/issues/1829)) ([5b6e4a2](https://github.com/Quantinuum/guppylang/commit/5b6e4a2ac2642ff66fa8ea091273c8d978189f6c))
+
+## [1.0.0-a3](https://github.com/Quantinuum/guppylang/compare/guppylang-internals-v1.0.0-a2...guppylang-internals-v1.0.0-a3) (2026-06-10)
+
+
+### ⚠ BREAKING CHANGES
+
+* Remove `power` flags and rename flag `dagger` to `daggerable` and `control` to `controllable` ([#1838](https://github.com/Quantinuum/guppylang/issues/1838))
+* Remove vacuous pytket extra from distributions ([#1836](https://github.com/Quantinuum/guppylang/issues/1836))
+* Experimental features should only be enabled globally ([#1825](https://github.com/Quantinuum/guppylang/issues/1825))
+* Error handling for outside use of variables scoped to modifier blocks ([#1712](https://github.com/Quantinuum/guppylang/issues/1712))
+
+### Features
+
+* Error handling for outside use of variables scoped to modifier blocks ([#1712](https://github.com/Quantinuum/guppylang/issues/1712)) ([cab8077](https://github.com/Quantinuum/guppylang/commit/cab8077732493ce120a8656b970af182e708bfac))
+
+
+### Bug Fixes
+
+* Allow check on generic functions that do not require monomorphisation ([#1666](https://github.com/Quantinuum/guppylang/issues/1666)) ([126124b](https://github.com/Quantinuum/guppylang/commit/126124bfdc8a566cc84d7727c472dc336380ed19))
+
+
+### Code Refactoring
+
+* Experimental features should only be enabled globally ([#1825](https://github.com/Quantinuum/guppylang/issues/1825)) ([5f01c7d](https://github.com/Quantinuum/guppylang/commit/5f01c7d6d2303bf3812496f8e91d20696b60e088))
+* Remove `power` flags and rename flag `dagger` to `daggerable` and `control` to `controllable` ([#1838](https://github.com/Quantinuum/guppylang/issues/1838)) ([e903bdb](https://github.com/Quantinuum/guppylang/commit/e903bdbafbe252bde224ea1c483e10180e265bfb))
+* Remove vacuous pytket extra from distributions ([#1836](https://github.com/Quantinuum/guppylang/issues/1836)) ([1fcb41f](https://github.com/Quantinuum/guppylang/commit/1fcb41fae1c8532cc624e648890badf8631c2159))
+
+## [1.0.0-a2](https://github.com/Quantinuum/guppylang/compare/guppylang-internals-v1.0.0-a1...guppylang-internals-v1.0.0-a2) (2026-06-08)
+
+
+### ⚠ BREAKING CHANGES
+
+* Make structs mutable by default ([#1795](https://github.com/Quantinuum/guppylang/issues/1795))
+* Higher order functions with unitary modifiers ([#1768](https://github.com/Quantinuum/guppylang/issues/1768))
+* Mutable structs ([#1793](https://github.com/Quantinuum/guppylang/issues/1793))
+* Raise an error if control flow is present in dagger contexts ([#1794](https://github.com/Quantinuum/guppylang/issues/1794))
+
+### Features
+
+* Higher order functions with unitary modifiers ([#1768](https://github.com/Quantinuum/guppylang/issues/1768)) ([fb6c58f](https://github.com/Quantinuum/guppylang/commit/fb6c58f1c5191c1b3fd722034f63edbc7c8143df))
+* Make structs mutable by default ([#1795](https://github.com/Quantinuum/guppylang/issues/1795)) ([6b20646](https://github.com/Quantinuum/guppylang/commit/6b206466cb1c1bb62d47943c64f05822ec739a34))
+* Mutable structs ([#1793](https://github.com/Quantinuum/guppylang/issues/1793)) ([06a551d](https://github.com/Quantinuum/guppylang/commit/06a551d4f8ab902f68e610e91ef0ea380996ae51))
+* Raise an error if control flow is present in dagger contexts ([#1794](https://github.com/Quantinuum/guppylang/issues/1794)) ([85e42d1](https://github.com/Quantinuum/guppylang/commit/85e42d15bf9d60dd384466f13919301300130316))
+
+
+### Bug Fixes
+
+* Comptime attribute access in struct definitions ([#1824](https://github.com/Quantinuum/guppylang/issues/1824)) ([89b39e2](https://github.com/Quantinuum/guppylang/commit/89b39e25459679b34baf2320abdf36e9371569f2))
+
 ## [1.0.0-a1](https://github.com/Quantinuum/guppylang/compare/guppylang-internals-v1.0.0-a0...guppylang-internals-v1.0.0-a1) (2026-06-04)